requirements.txt
appnope==0.1.2
asgiref==3.3.1
backcall==0.2.0
bcrypt==3.2.0
boto3==1.16.63
botocore==1.19.63
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
decorator==4.4.2
Django==3.1.5
django-cors-headers==3.6.0
django-seed==0.2.2
Faker==5.8.0
gunicorn==20.0.4
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
jmespath==0.10.0
mysqlclient==2.0.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.14
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.7.4
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
